The full commit message for this commit is long. Read in full.
The material specific code is separated completely from the main problem
code file. A new file called viscoelastic_mechanics_bottom_up.h was
created. The intention is to use this file as a basis for building a
complex, finite deformation viscoelastic model from the bottom up. The
starting point will be a small deformation elastic model, followed by a
small deformation viscoelastic model with a single viscous element. Then
that will be modified to a small deformation viscoelastic model with two
elements and the final version will be a large deformation viscoelastic
similar to the model in "Biomechanics of brain tissue" by Prevost et.
al. published in 2011 in Acta Biomaterialia.